+++ ActiveField.php
@@ -385,8 +385,8 @@ class ActiveField extends \yii\widgets\ActiveField
             $config['horizontalCssClasses'] = $cssClasses;
             $config['wrapperOptions'] = ['class' => $cssClasses['wrapper']];
             $config['labelOptions'] = ['class' => 'control-label ' . $cssClasses['label']];
-            $config['errorOptions'] = ['class' => 'help-block help-block-error ' . $cssClasses['error']];
-            $config['hintOptions'] = ['class' => 'help-block ' . $cssClasses['hint']];
+            $config['errorOptions']['class'] = 'help-block help-block-error ' . $cssClasses['error'];
+            $config['hintOptions']['class'] = 'help-block ' . $cssClasses['hint'];
         } elseif ($layout === 'inline') {
             $config['labelOptions'] = ['class' => 'sr-only'];
             $config['enableError'] = false;